+++ include_public/avsystem/commons/avs_errno_map.h
@@ -23,9 +23,9 @@
 extern "C" {
 #endif
 
-// Note: It might seem more appropriate to write #ifndef errno, but even though
-// ISO C defines errno as being a macro, on some platforms (e.g. lwIP) it isn't
-// a macro, but rather just an external linkage symbol.
+// Note: It might seem more appropriate to write #ifndef errno, but wording
+// seems to allow to errno to be not a macro, but an external linkage symbol and
+// lwIP does just that.
 #ifndef EDOM
 #    error "For this header to be useful, you have to include your system / library / whatever errno.h first."
 #endif // EDOM